requirements.txt
pyspark
pandas
matplotlib
python-dotenv
pytest>=7.0.0
requests>=2.28.0
pytest-mock>=3.10.0
pytest-html
great-expectations>=0.15.0
azure-storage-blob>=12.0.0
pyarrow>=7.0.0
pytest-dependency>=0.5.1
